move test helpers that need testing pkg to _test.go file
